requirements.txt
numpy>=1.18.0
torch>=1.7.0
matplotlib>=3.2.0
pyyaml>=5.3.1
scikit-learn>=0.24.0
tqdm>=4.50.0
jupyterlab>=3.0.0
